added signIn to the containers and added navBar
