+++ src/client/App.tsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import { Navigate, Routes, Route } from 'react-router-dom';
 import Display from './containers/Display';
+import SignIn from './containers/SignIn';
 
 function App() {
   return (
-    <div id='mainApp'>
-      <Display />
+    <div id='mainApp' className='main'>
+      <SignIn />
+      {/* <Display /> */}
       {/* <Routes>
         <Route path="/Display/*" element={<Display />} />
         <Route path="/" element={<Setup />} />
