+++ src/client/components/NavBar.tsx
@@ -13,8 +13,14 @@ import MailIcon from '@mui/icons-material/Mail';
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import MoreIcon from '@mui/icons-material/MoreVert';
 const Icon = require('../../../public/HWIcon.svg').default;
+import { useSelector, useDispatch } from 'react-redux';
+import type { HireWallState, ApplicationRecord } from '../slice';
+import { setApplicationRecords, setActionRecords } from '../slice';
 
 export default function NavBar() {
+  const isLoggedIn = useSelector((state: HireWallState) => state.isLoggedIn);
+  const user = useSelector((state: HireWallState) => state.userRecord);
+
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
   const [mobileMoreAnchorEl, setMobileMoreAnchorEl] =
     React.useState<null | HTMLElement>(null);
@@ -56,8 +62,9 @@ export default function NavBar() {
       open={isMenuOpen}
       onClose={handleMenuClose}
     >
-      <MenuItem onClick={handleMenuClose}>Profile</MenuItem>
-      <MenuItem onClick={handleMenuClose}>My account</MenuItem>
+      <MenuItem onClick={handleMenuClose}>
+        {isLoggedIn ? 'Log Out' : 'Log In'}
+      </MenuItem>
     </Menu>
   );
 
@@ -78,15 +85,15 @@ export default function NavBar() {
       open={isMobileMenuOpen}
       onClose={handleMobileMenuClose}
     >
-      <MenuItem>
+      {/* <MenuItem>
         <IconButton size='large' aria-label='show 4 new mails' color='inherit'>
           <Badge badgeContent={4} color='error'>
             <MailIcon />
           </Badge>
         </IconButton>
         <p>Messages</p>
-      </MenuItem>
-      <MenuItem>
+      </MenuItem> */}
+      {/* <MenuItem>
         <IconButton
           size='large'
           aria-label='show 17 new notifications'
@@ -97,7 +104,7 @@ export default function NavBar() {
           </Badge>
         </IconButton>
         <p>Notifications</p>
-      </MenuItem>
+      </MenuItem> */}
       <MenuItem onClick={handleProfileMenuOpen}>
         <IconButton
           size='large'
@@ -108,7 +115,7 @@ export default function NavBar() {
         >
           <AccountCircle />
         </IconButton>
-        <p>Profile</p>
+        <p>{user.name === '' ? 'LogIn' : user.name}</p>
       </MenuItem>
     </Menu>
   );
@@ -138,7 +145,7 @@ export default function NavBar() {
 
           <Box sx={{ flexGrow: 1 }} />
           <Box sx={{ display: { xs: 'none', md: 'flex' } }}>
-            <IconButton
+            {/* <IconButton
               size='large'
               aria-label='show 4 new mails'
               color='inherit'
@@ -155,7 +162,7 @@ export default function NavBar() {
               <Badge badgeContent={17} color='error'>
                 <NotificationsIcon />
               </Badge>
-            </IconButton>
+            </IconButton> */}
             <IconButton
               size='large'
               edge='end'
