+++ src/client/containers/Display.tsx
@@ -8,7 +8,11 @@ import FilterButton from '../components/FilterButton';
 import ViewButton from '../components/ViewButton';
 import { useSelector, useDispatch } from 'react-redux';
 import type { HireWallState, ApplicationRecord } from '../slice';
-import { setApplicationRecords, setActionRecords } from '../slice';
+import {
+  setApplicationRecords,
+  setActionRecords,
+  setIsLoggedIn,
+} from '../slice';
 
 function Display() {
   const [allowList, setAllowList] = useState<Set<number>>(new Set());
@@ -31,17 +35,19 @@ function Display() {
   }, [searchTerm]);
 
   useEffect(() => {
+    dispatch(setIsLoggedIn(true));
+  }, []);
+
+  useEffect(() => {
     //insert fetch request to back end for app records
     // Assume that the user record has been setup by the LOGIN Page
-
     // need to POST with user_id
-    fetch('/application', {
-      method: 'POST',
+    fetch(`/application`, {
+      method: 'GET',
       headers: {
         Accept: 'application/json',
         'Content-Type': 'application/json',
       },
-      body: JSON.stringify({ user_id: user._id }), //search by user id
     })
       .then((res) => res.json())
       .then((data) => {
