+++ src/client/styles/index.scss
@@ -61,3 +61,15 @@
   justify-content: center;
   flex-wrap: wrap;
 }
+
+.center {
+  display: flex;
+  direction: column;
+  justify-content: center;
+}
+
+.signIn {
+  display: flex;
+  direction: column;
+  justify-content: center;
+}
